+++ Fluent.Ribbon/Controls/GalleryGroupContainer.cs
@@ -1,13 +1,9 @@
 // ReSharper disable once CheckNamespace
 namespace Fluent
 {
-    using System;
-    using System.Diagnostics;
     using System.Windows;
     using System.Windows.Controls;
     using System.Windows.Media;
-    using Fluent.Extensions;
-    using Fluent.Internal;
     using Fluent.Internal.KnownBoxes;
 
     /// <summary>
@@ -15,17 +11,6 @@ namespace Fluent
     /// </summary>
     public class GalleryGroupContainer : HeaderedItemsControl
     {
-        #region Fields
-
-        private int previousItemsCount;
-
-        // Whether MinWidth/MaxWidth of the ItemsPanel needs to be updated
-        private bool minMaxWidthNeedsToBeUpdated = true;
-        private Panel itemsPanel;
-        private FrameworkElement targetForSizeConstraints;
-
-        #endregion
-
         #region Properites
 
         #region IsHeadered
@@ -127,12 +112,7 @@ namespace Fluent
         /// </summary>
         public static readonly DependencyProperty MinItemsInRowProperty =
             DependencyProperty.Register(nameof(MinItemsInRow), typeof(int),
-            typeof(GalleryGroupContainer), new FrameworkPropertyMetadata(IntBoxes.Zero, FrameworkPropertyMetadataOptions.AffectsMeasure, OnMinItemsInRowChanged));
-
-        private static void OnMinItemsInRowChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
-        {
-            OnMaxOrMinItemsInRowChanged(d, e);
-        }
+            typeof(GalleryGroupContainer), new FrameworkPropertyMetadata(IntBoxes.Zero, FrameworkPropertyMetadataOptions.AffectsMeasure));
 
         #endregion
 
@@ -150,34 +130,13 @@ namespace Fluent
         /// <summary>
         /// <see cref="DependencyProperty"/> for <see cref="MaxItemsInRow"/>.
         /// </summary>
-        public static readonly DependencyProperty MaxItemsInRowProperty = DependencyProperty.Register(nameof(MaxItemsInRow), typeof(int), typeof(GalleryGroupContainer), new FrameworkPropertyMetadata(0, FrameworkPropertyMetadataOptions.AffectsMeasure, OnMaxItemsInRowChanged));
-
-        private static void OnMaxItemsInRowChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
-        {
-            OnMaxOrMinItemsInRowChanged(d, e);
-        }
+        public static readonly DependencyProperty MaxItemsInRowProperty = DependencyProperty.Register(nameof(MaxItemsInRow), typeof(int), typeof(GalleryGroupContainer), new FrameworkPropertyMetadata(0, FrameworkPropertyMetadataOptions.AffectsMeasure));
 
         #endregion
-
-        private Panel RealItemsPanel => this.itemsPanel ?? (this.itemsPanel = FindItemsPanel(this));
-
-        private static void OnMaxOrMinItemsInRowChanged(DependencyObject d, DependencyPropertyChangedEventArgs e)
-        {
-            var galleryGroupContainer = (GalleryGroupContainer)d;
-            galleryGroupContainer.minMaxWidthNeedsToBeUpdated = true;
-        }
 
         #endregion
 
         #region Initialization
-
-        /// <summary>
-        /// Creates a new instance of <see cref="GalleryGroupContainer"/>.
-        /// </summary>
-        public GalleryGroupContainer()
-        {
-            this.Unloaded += this.HandleUnloaded;
-        }
 
         /// <summary>
         /// Static constructor
@@ -192,163 +151,9 @@ namespace Fluent
         {
             base.OnItemsPanelChanged(oldItemsPanel, newItemsPanel);
 
-            this.itemsPanel = null;
-            this.minMaxWidthNeedsToBeUpdated = true;
             this.InvalidateMeasure();
         }
 
         #endregion
-
-        #region MaxWidth Updating
-
-        // Sets MaxWidth of the items panel based of ItemsInRow property
-        private void UpdateMinAndMaxWidth(Size constraint)
-        {
-            if (this.minMaxWidthNeedsToBeUpdated == false)
-            {
-                return;
-            }
-
-            if (this.RealItemsPanel == null)
-            {
-                // Item's panel is not ready now
-                if (this.IsLoaded)
-                {
-                    Debug.WriteLine("Panel with IsItemsHost = true is not found in GalleryGroupContainer (probably the style is not correct or haven't attached yet)");
-                }
-                else
-                {
-                    // Prevent duplicate registration
-                    this.Loaded -= this.HandleLoaded;
-                    this.Loaded += this.HandleLoaded;
-                }
-
-                return;
-            }
-
-            this.minMaxWidthNeedsToBeUpdated = false;
-
-            // Issue references:
-            // - #542 + commit https://github.com/fluentribbon/Fluent.Ribbon/commit/8b458b1cfc5e440f54778c808142fffa67a23978
-            // - #666
-            // We need to check if we are inside a closed InRibbonGallery.
-            // - If we are inside an closed InRibbonGallery we need to restrict the size of "this"
-            // - If we are inside an opened InRibbonGallery or not inside an InRibbonGallery we need to restrict the size of "RealItemsPanel"
-            var inRibbonGallery = UIHelper.GetParent<InRibbonGallery>(this);
-            var isInsideClosedInRibbonGallery = inRibbonGallery != null && inRibbonGallery.IsDropDownOpen == false;
-            this.targetForSizeConstraints = isInsideClosedInRibbonGallery ? (FrameworkElement)this : this.RealItemsPanel;
-
-            var nonTargetForSizeConstraints = isInsideClosedInRibbonGallery ? (FrameworkElement)this.RealItemsPanel : this;
-
-            nonTargetForSizeConstraints.MinWidth = 0;
-            nonTargetForSizeConstraints.MaxWidth = double.PositiveInfinity;
-
-            if (this.Orientation == Orientation.Vertical)
-            {
-                // Min/Max is used for Horizontal layout only
-                this.targetForSizeConstraints.MinWidth = 0;
-                this.targetForSizeConstraints.MaxWidth = double.PositiveInfinity;
-                return;
-            }
-
-            var itemWidth = this.GetItemSize().Width;
-            if (double.IsNaN(itemWidth))
-            {
-                // We can't calc item's width now
-                return;
-            }
-
-            this.targetForSizeConstraints.MinWidth = (Math.Min(this.Items.Count, this.MinItemsInRow) * itemWidth) + 0.1;
-            this.targetForSizeConstraints.MaxWidth = this.MaxItemsInRow == 0
-                ? constraint.Width
-                : (Math.Min(this.Items.Count, this.MaxItemsInRow) * itemWidth) + 0.1;
-        }
-
-        private void HandleLoaded(object sender, RoutedEventArgs e)
-        {
-            this.Loaded -= this.HandleLoaded;
-
-            if (this.minMaxWidthNeedsToBeUpdated == false)
-            {
-                return;
-            }
-
-            this.InvalidateMeasure();
-        }
-
-        private void HandleUnloaded(object sender, RoutedEventArgs e)
-        {
-            this.itemsPanel = null;
-
-            this.minMaxWidthNeedsToBeUpdated = true;
-        }
-
-        /// <summary>
-        /// Determines the desired item size.
-        /// </summary>
-        /// <returns>
-        /// <see cref="Size"/> constructed from <see cref="ItemWidth"/> and <see cref="ItemHeight"/>.
-        /// If no items are present <see cref="Size.Empty"/>.
-        /// </returns>
-        public Size GetItemSize()
-        {
-            if (this.Items.Count == 0)
-            {
-                return Size.Empty;
-            }
-
-            return new Size(this.ItemWidth, this.ItemHeight);
-        }
-
-        // Finds panel with IsItemsHost, or null if such panel is not found
-        private static Panel FindItemsPanel(DependencyObject obj)
-        {
-            for (var i = 0; i < VisualTreeHelper.GetChildrenCount(obj); i++)
-            {
-                if (obj is Panel panel &&
-                    panel.IsItemsHost)
-                {
-                    return panel;
-                }
-
-                panel = FindItemsPanel(VisualTreeHelper.GetChild(obj, i));
-                if (panel != null)
-                {
-                    return panel;
-                }
-            }
-
-            return null;
-        }
-
-        #endregion
-
-        /// <inheritdoc />
-        protected override Size MeasureOverride(Size constraint)
-        {
-            if (this.previousItemsCount != this.Items.Count
-                || this.minMaxWidthNeedsToBeUpdated)
-            {
-                // Track ItemsPanel changes
-                this.previousItemsCount = this.Items.Count;
-                this.minMaxWidthNeedsToBeUpdated = true;
-
-                this.UpdateMinAndMaxWidth(constraint);
-            }
-
-            if (this.targetForSizeConstraints != null)
-            {
-                return base.MeasureOverride(new Size(this.targetForSizeConstraints.MaxWidth, constraint.Height));
-            }
-
-            return base.MeasureOverride(constraint);
-            // todo: maybe the following code is a better alternative
-            //if (this.targetForSizeConstraints != null)
-            //{
-            //    return base.MeasureOverride(new Size(Math.Min(constraint.Width, this.targetForSizeConstraints.MaxWidth), constraint.Height));
-            //}
-
-            //return Size.Empty;
-        }
     }
 }