+++ Changelog.md
@@ -5,6 +5,7 @@
   - [#786](../../issues/786) - InRibbonGallery: Dynamic ItemWidth / use translated Text in items
   - [#788](../../issues/788) - Maximize icon is incorrectly drawn on high-dpi displays
   - [#789](../../issues/789) - Opening gallery messes up InRibbonGallery resizing.
+  - [#791](../../issues/791) - Problem with dynamic items in InRibbonGallery
 
 ## 7.1.0
 - ### Bug fixes